requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
google-cloud-storage==1.42.3
ffmpeg-python==0.2.0
functions-framework==3.0.0
